Add sanitizers module
This module contains the implementation of all hooks added by gofuzz.
Therefore, the module must be added as a dependency to the tested code.
This commit includes a sample hook to check if we can execute arbitrary
commands and try guiding the fuzzer to execute an evil command we
control.